+++ packages/extension-polkagate/src/components/AddressInput.tsx
@@ -28,10 +28,11 @@ interface Props {
   style?: SxProps<Theme>;
   withSelect?: boolean;
   setIsError?: React.Dispatch<React.SetStateAction<boolean | undefined>>;
-  setType?: React.Dispatch<React.SetStateAction<KeypairType | undefined>>
+  setType?: React.Dispatch<React.SetStateAction<KeypairType | undefined>>;
+  showAddressBook?: boolean;
 }
 
-export default function AddressInput({ addWithQr = false, address, chain, disabled = false, label, placeHolder, setAddress, setIsError, setType, style, withSelect }: Props): React.ReactElement<Props> {
+export default function AddressInput({ addWithQr = false, address, chain, disabled = false, label, placeHolder, setAddress, setType, setIsError, showAddressBook, style, withSelect }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
   const containerRef = useRef<HTMLDivElement>(null);
   const isBlueish = useIsBlueish();
@@ -214,6 +215,7 @@ export default function AddressInput({ addWithQr = false, address, chain, disabl
           handleClose={() => setOpenAccountList(false)}
           open={openAccountList}
           setAddress={setAddress}
+          showAddressBook={showAddressBook}
         />
       }
     </>