+++ packages/extension-polkagate/src/fullscreen/settings/AccountSettings.tsx
@@ -1,11 +1,12 @@
 // Copyright 2019-2026 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { Stack, Typography } from '@mui/material';
-import { Broom, Edit2, ExportCurve, type Icon, ImportCurve, LogoutCurve, Notification as NotificationIcon, ShieldSecurity } from 'iconsax-react';
+import { Container, Stack, Typography } from '@mui/material';
+import { Book, Broom, Data2, Edit2, ExportCurve, type Icon, ImportCurve, LogoutCurve, Notification as NotificationIcon, ShieldSecurity } from 'iconsax-react';
 import React, { useCallback, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 
+import useIsHovered from '@polkadot/extension-polkagate/src/hooks/useIsHovered2';
 import RemoveAccount from '@polkadot/extension-polkagate/src/partials/RemoveAccount';
 import { ExtensionPopups } from '@polkadot/extension-polkagate/src/util/constants';
 import { useExtensionPopups } from '@polkadot/extension-polkagate/src/util/handleExtensionPopup';
@@ -18,6 +19,8 @@ import DeriveAccount from '../home/DeriveAccount';
 import ExportAllAccounts from '../home/ExportAllAccounts';
 import RenameAccount from '../home/RenameAccount';
 import NotificationSettingsFS from '../notification/NotificationSettingsFS';
+import ProxiedAccount from './importProxied/ProxiedAccount';
+import AddressBook from './addressBook';
 
 interface ActionBoxProps {
   Icon: Icon;
@@ -28,6 +31,7 @@ interface ActionBoxProps {
 
 function ActionBox({ Icon, label, onClick, path }: ActionBoxProps): React.ReactElement {
   const navigate = useNavigate();
+ const { isHovered, ref } = useIsHovered();
 
   const _onClick = useCallback(() => {
     onClick
@@ -36,8 +40,8 @@ function ActionBox({ Icon, label, onClick, path }: ActionBoxProps): React.ReactE
   }, [navigate, onClick, path]);
 
   return (
-    <Stack direction='column' justifyContent='start' onClick={_onClick} rowGap='8px' sx={{ '&:hover': { bgcolor: '#2D1E4A', transform: 'translateY(-4px)' }, bgcolor: '#05091C', borderRadius: '14px', cursor: 'pointer', height: '86px', minWidth: '141px', px: '10px', transition: 'all 250ms ease-out', width: 'fit-content' }}>
-      <Icon color='#AA83DC' size='24' style={{ marginTop: '17px' }} variant='Bulk' />
+    <Stack direction='column' justifyContent='start' onClick={_onClick} ref={ref} rowGap='8px' sx={{ '&:hover': { bgcolor: '#2D1E4A', transform: 'translateY(-4px)' }, bgcolor: '#05091C', borderRadius: '14px', cursor: 'pointer', height: '86px', minWidth: '141px', px: '10px', transition: 'all 250ms ease-out', width: 'fit-content' }}>
+      <Icon color='#AA83DC' size='24' style={{ marginTop: '17px' }} variant={isHovered ? 'Bold' : 'Bulk'} />
       <Typography sx={{ display: 'flex', fontWeight: 700, width: 'fit-content' }} variant='B-2'>
         {label}
       </Typography>
@@ -97,6 +101,17 @@ function AccountSettings(): React.ReactElement {
           />
         );
 
+      case ExtensionPopups.IMPORT_PROXIED:
+        return (
+          <ProxiedAccount closePopup={extensionPopupCloser} mode='import' />
+        );
+
+      case ExtensionPopups.ADDRESS_BOOK:
+        return (
+          <AddressBook
+            closePopup={extensionPopupCloser}
+          />);
+
       default:
         return null;
     }
@@ -109,38 +124,58 @@ function AccountSettings(): React.ReactElement {
           <Typography color='text.primary' fontSize='22px' mt='30px' sx={{ display: 'block', textAlign: 'left', textTransform: 'uppercase' }} variant='H-4'>
             {t('Actions')}
           </Typography>
-          <VelvetBox childrenStyle={{ columnGap: '4px', display: 'flex', flexDirection: 'row' }} style={{ margin: '20px 0', width: 'fit-content' }}>
-            <ActionBox
-              Icon={NotificationIcon}
-              label={t('Notification')}
-              onClick={extensionPopupOpener(ExtensionPopups.NOTIFICATION)}
-            />
-            <ActionBox
-              Icon={Edit2}
-              label={t('Rename Account')}
-              onClick={extensionPopupOpener(ExtensionPopups.RENAME)}
-            />
-            <ActionBox
-              Icon={ExportCurve}
-              label={t('Export Accounts')}
-              onClick={extensionPopupOpener(ExtensionPopups.EXPORT)}
-            />
-            <ActionBox
-              Icon={ImportCurve}
-              label={t('Import Accounts')}
-              path='/account/have-wallet'
-            />
-            <ActionBox
-              Icon={ShieldSecurity}
-              label={t('Manage Website Access')}
-              onClick={extensionPopupOpener(ExtensionPopups.DAPPS)}
-            />
-            {!selectedAccount?.isExternal &&
+          <VelvetBox style={{ margin: 0, marginTop: '20px', width: 'fit-content' }}>
+            <Container style={{ columnGap: '4px', display: 'flex', flexDirection: 'row', margin: 0, padding: 0, width: 'fit-content' }}>
               <ActionBox
-                Icon={Broom}
-                label={t('Derive from Account')}
-                onClick={extensionPopupOpener(ExtensionPopups.DERIVE)}
-              />}
+                Icon={Edit2}
+                label={t('Rename Account')}
+                onClick={extensionPopupOpener(ExtensionPopups.RENAME)}
+              />
+              <ActionBox
+                Icon={ExportCurve}
+                label={t('Export Accounts')}
+                onClick={extensionPopupOpener(ExtensionPopups.EXPORT)}
+              />
+              <ActionBox
+                Icon={ImportCurve}
+                label={t('Import Accounts')}
+                path='/account/have-wallet'
+              />
+              {!selectedAccount?.isExternal &&
+                <ActionBox
+                  Icon={Broom}
+                  label={t('Derive from Account')}
+                  onClick={extensionPopupOpener(ExtensionPopups.DERIVE)}
+                />}
+            </Container>
+          </VelvetBox>
+          <Typography color='text.primary' fontSize='22px' mt='30px' sx={{ display: 'block', textAlign: 'left', textTransform: 'uppercase' }} variant='H-4'>
+            {t('Options')}
+          </Typography>
+          <VelvetBox style={{ margin: 0, marginTop: '15px', width: 'fit-content' }}>
+            <Container style={{ columnGap: '4px', display: 'flex', flexDirection: 'row', margin: 0, padding: 0, width: 'fit-content' }}>
+              <ActionBox
+                Icon={Book}
+                label={t('Address Book')}
+                onClick={extensionPopupOpener(ExtensionPopups.ADDRESS_BOOK)}
+              />
+              <ActionBox
+                Icon={Data2}
+                label={t('Import Proxied')}
+                onClick={extensionPopupOpener(ExtensionPopups.IMPORT_PROXIED)}
+              />
+              <ActionBox
+                Icon={NotificationIcon}
+                label={t('Notification')}
+                onClick={extensionPopupOpener(ExtensionPopups.NOTIFICATION)}
+              />
+              <ActionBox
+                Icon={ShieldSecurity}
+                label={t('Manage Website Access')}
+                onClick={extensionPopupOpener(ExtensionPopups.DAPPS)}
+              />
+
+            </Container>
           </VelvetBox>
           <Stack alignItems='center' columnGap='5px' direction='row' onClick={extensionPopupOpener(ExtensionPopups.REMOVE)} sx={{ bottom: '20px', cursor: 'pointer', position: 'absolute' }}>
             <LogoutCurve color='#AA83DC' size={18} variant='Bulk' />